+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/muir/gwrap
 
-go 1.18
+go 1.20
 
 require (
 	github.com/stretchr/testify v1.10.0
